Encode all program info types
