+++ libs/api_common/api_common.vcxproj
@@ -162,6 +162,7 @@
     <ClInclude Include="tlv.h" />
     <ClInclude Include="device_helper.hpp" />
     <ClInclude Include="windows_platform_common.hpp" />
+    <ClInclude Include="windows_program_type.h" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\tools\encode_program_info\encode_program_info.vcxproj">
