+++ libs/api_common/api_common.vcxproj.filters
@@ -48,5 +48,8 @@
     <ClInclude Include="device_helper.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="windows_program_type.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
