+++ tools/encode_program_info/encode_program_info.vcxproj.filters
@@ -23,4 +23,7 @@
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
+  <ItemGroup>
+    <None Include="packages.config" />
+  </ItemGroup>
 </Project>
