+++ tools/encode_program_info/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="boost" version="1.75.0.0" targetFramework="native" />
+</packages>
